@@ -10,10 +10,10 @@
     "bootstrap": "^4.5.2",
     "chart.js": "^3.7.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
+    "react": "^18.0.0-rc.2",
     "react-bootstrap": "^1.4.0",
     "react-chartjs-2": "^4.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.0.0-rc.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0",
     "sass": "^1.26.10",
