@@ -1,16 +1,15 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import 'bootstrap/dist/css/bootstrap.css'
 import 'wowjs/css/libs/animate.css'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
+  <App />
 )
 
 // If you want to start measuring performance in your app, pass a function
